refactor: inversao de dependencia no modulo de clientes

+++ api/src/modules/clients/services/clients.service.ts
@@ -1,33 +1,26 @@
 import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import { IValidateUserOwnershipService } from 'src/modules/users/interfaces/validate-user-ownership-service.interface';
-import { ClientsRespository } from 'src/shared/database/repositories/clients.repository';
+import { IClientsRepository } from 'src/shared/database/interfaces/clients-repository.interface';
 import { CreateClientDto } from '../dto/create-client.dto';
 import { UpdateClientDto } from '../dto/update-client.dto';
-import { ValidateClientOwnershipService } from './validate-client-ownership.service';
+import { IClientsService } from '../interfaces/clients-service.interface';
+import { IValidateClientOwnershipService } from '../interfaces/validate-client-ownership-service.interface';
 
 @Injectable()
-export class ClientsService {
+export class ClientsService implements IClientsService {
   constructor(
-    private readonly clientsRepository: ClientsRespository,
+    @Inject(IClientsRepository)
+    private readonly clientsRepository: IClientsRepository,
     @Inject(IValidateUserOwnershipService)
     private readonly validateUserOwnershipService: IValidateUserOwnershipService,
-    private readonly validateClientOwnershipService: ValidateClientOwnershipService,
+    @Inject(IValidateClientOwnershipService)
+    private readonly validateClientOwnershipService: IValidateClientOwnershipService,
   ) {}
 
   async findAllByUserId(userId: string) {
-    const clients = await this.clientsRepository.findMany({
-      where: {
-        userId,
-      },
-      select: {
-        id: true,
-        name: true,
-        phone: true,
-        type: true,
-        address: true,
-        document: true,
-        balance: true,
-      },
+    const clients = await this.clientsRepository.findManyByUserId({
+      userId,
+      order: 'asc',
     });
 
     const sortedClients = clients.sort((a, b) =>
@@ -38,21 +31,7 @@ export class ClientsService {
   }
 
   async findOneByUserId(userId: string, clientId: string) {
-    return this.clientsRepository.findFirst({
-      where: {
-        userId,
-        id: clientId,
-      },
-      select: {
-        id: true,
-        name: true,
-        phone: true,
-        type: true,
-        address: true,
-        document: true,
-        balance: true,
-      },
-    });
+    return this.clientsRepository.findFirstById({ userId, id: clientId });
   }
 
   async create(userId: string, createClientDto: CreateClientDto) {
@@ -62,12 +41,8 @@ export class ClientsService {
       createClientDto;
 
     if (document) {
-      const documentAlreadyExists = await this.clientsRepository.findFirst({
-        where: { userId, document },
-        select: {
-          document: true,
-        },
-      });
+      const documentAlreadyExists =
+        await this.clientsRepository.findFirstByDocument({ userId, document });
 
       const errorMessage =
         type === 'FISICO' ? 'CPF já cadastrado.' : 'CNPJ já cadastrado.';
@@ -78,12 +53,12 @@ export class ClientsService {
     }
 
     return this.clientsRepository.create({
+      userId,
       data: {
-        userId,
         name,
-        phone: phone || null,
-        address: address || null,
-        document: document || null,
+        phone,
+        address,
+        document,
         type,
         balance: initialBalance,
       },
@@ -100,12 +75,9 @@ export class ClientsService {
     const { name, phone, address, document, type, balance } = updateClientDto;
 
     if (document) {
-      const client = await this.clientsRepository.findFirst({
-        where: { userId, document },
-        select: {
-          document: true,
-          id: true,
-        },
+      const client = await this.clientsRepository.findFirstByDocument({
+        userId,
+        document,
       });
 
       const errorMessage =
@@ -117,23 +89,15 @@ export class ClientsService {
     }
 
     return this.clientsRepository.update({
-      where: { userId, id: clientId },
+      userId,
       data: {
+        id: clientId,
         name,
-        phone: phone || null,
-        address: address || null,
-        document: document || null,
+        phone,
+        address,
+        document,
         type,
         balance,
-      },
-      select: {
-        id: true,
-        name: true,
-        phone: true,
-        address: true,
-        document: true,
-        type: true,
-        balance: true,
       },
     });
   }
@@ -141,9 +105,7 @@ export class ClientsService {
   async remove(userId: string, clientId: string) {
     await this.validateClientOwnershipService.validate(userId, clientId);
 
-    await this.clientsRepository.delete({
-      where: { id: clientId },
-    });
+    await this.clientsRepository.delete({ userId, id: clientId });
 
     return null;
   }
